feat: Merge | CleanUp layout
