+++ src/application/Controller.java
@@ -1,3 +1,4 @@
+package application;
 
 /**
  * This class represents the controller for the application
@@ -11,7 +12,7 @@ public class Controller {
 
 	public Controller() {
 		
-		// TODO: dragging model.cards from one column to another
+		// TODO: dragging application.model.cards from one column to another
 		
 	}
 }
