+++ src/application/Main.java
@@ -1,3 +1,6 @@
+package application;
+
+import application.model.Model;
 import javafx.application.Application;
 import javafx.stage.Stage;
 import javafx.scene.Scene;
@@ -14,7 +17,6 @@ import javafx.scene.Scene;
 public class Main extends Application {
 
   public static final Model model = new Model();
-  public static final BoardView boardView = new BoardView();
   public static final SplashView splashView = new SplashView();
   public static final Controller controller = new Controller();
