+++ src/application/model/cards/Card.java
@@ -1,9 +1,9 @@
-package model.cards;
+package application.model.cards;
 
 import javafx.beans.property.SimpleStringProperty;
 
 /**
- * This class represents the Card Object which has a title
+ * This class represents the application.Card Object which has a title
  *
  * @author Jason Cleveland
  * @author Joshua Laver
